🧪 [testing improvement] Add tests for synchronous project-settings utilities
🎯 What: Moved existing sync tests to a new isolated file tests/helpers/project-settings-sync.test.ts, and added tests for error cases in parseProjectSettings.
📊 Coverage: parseProjectSettings now handles readFileSync errors properly with 100% path coverage.
✨ Result: Improved test coverage and isolation for synchronous settings parser logic.
