+++ tests/helpers/project-settings.test.ts
@@ -2,22 +2,14 @@
  * Tests for project.godot settings parser and manipulation
  */
 
-import * as fs from 'node:fs'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import {
   getInputActions,
   getSetting,
-  parseProjectSettings,
   parseProjectSettingsContent,
   setSettingInContent,
-  writeProjectSettings,
 } from '../../src/tools/helpers/project-settings.js'
 import { SAMPLE_PROJECT_GODOT } from '../fixtures.js'
-
-vi.mock('node:fs', () => ({
-  readFileSync: vi.fn(),
-  writeFileSync: vi.fn(),
-}))
 
 describe('project-settings', () => {
   beforeEach(() => {
@@ -191,35 +183,6 @@ describe('project-settings', () => {
       const settings = parseProjectSettingsContent('[application]\nconfig/name="Test"\n')
       const actions = getInputActions(settings)
       expect(actions.size).toBe(0)
-    })
-  })
-
-  // ==========================================
-  // parseProjectSettings
-  // ==========================================
-  describe('parseProjectSettings', () => {
-    it('should parse project settings synchronously', () => {
-      const mockContent = '[application]\nconfig/name="Test"'
-      vi.mocked(fs.readFileSync).mockReturnValue(mockContent)
-
-      const settings = parseProjectSettings('project.godot')
-
-      expect(fs.readFileSync).toHaveBeenCalledWith('project.godot', 'utf-8')
-      expect(settings.sections.get('application')?.get('config/name')).toBe('"Test"')
-    })
-  })
-
-  // ==========================================
-  // writeProjectSettings
-  // ==========================================
-  describe('writeProjectSettings', () => {
-    it('should write project settings synchronously', () => {
-      const mockContent = 'some content'
-      vi.mocked(fs.writeFileSync).mockReturnValue(undefined)
-
-      writeProjectSettings('project.godot', mockContent)
-
-      expect(fs.writeFileSync).toHaveBeenCalledWith('project.godot', mockContent, 'utf-8')
     })
   })
 })
